adding tests for audit flow
