+++ arlo-client/src/App.css
@@ -3,7 +3,6 @@
 html {
   box-sizing: border-box;
   line-height: 1;
-  user-select: none;
 }
 *,
 *::before,
